Struct rsml::random_forest::model::RandomForest

pub struct RandomForest {
    pub trees: Vec<DecisionTree>,
}

This represents the Random Forest.

Fields

trees

Vector of Decision Trees

Methods

impl RandomForest

fn new(n_estimators: usize) -> RandomForest

Create a new random forest

Arguments:

n_estimators - number of decision trees you want

fn bootstrap(num_samples: usize) -> Vec<usize>

get a random set of indices to map random samples to a decision tree in your forest

Arguments:

num_indices - number of samples in the training data