Struct rsml::tree::model::DecisionTree

pub struct DecisionTree {
    // some fields omitted
}

DecisionTree represents the full decision tree model

Methods

impl DecisionTree

fn new() -> DecisionTree

create new decision tree

fn split_data(X: &Mat<f64>, indices: &Col<usize>, feature_idx: usize, threshold: f64) -> (Col<usize>, Col<usize>)

Split a feature column in training data.

Arguments:

X - original training data

indices - indices of data subset fed to node

feature_idx - index of current feature to split on

threshold - threshold to apply to current split

Example:

let X = RcArray::random((10,5), Range::new(0.,10.));
let indices = RcArray::from_vec(vec![0,1,2,3,4,5,6,7,8,9]);
let feature_idx = 4;
let value = 4.0;
let (left, right) = DecisionTree::split_data(&X,&indices, feature_idx, value);
assert!(left.iter().all(|&x| X.get((x,feature_idx)).unwrap() <= &value));
assert!(right.iter().all(|&x| X.get((x,feature_idx)).unwrap() > &value));

fn calculate_split(X: &Mat<f64>, feature_idx: usize, y: &Col<f64>, indices: &Col<usize>) -> (f64, f64)

Determine optimal threshold to split data

Arguments:

X - original training data

feature_idx - index of current feature to split on

y - original target data

indices - indices of data subset fed to node

Example:

let X = rcarr2(&[[-1.0], [-0.5], [0.0], [0.0],[0.0],[0.5],[1.0]]);
let y = RcArray::from_vec(vec![1.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0]);
let indices = RcArray::from_vec(vec![0,1,2,3,4,5,6]);
let (threshold, split_impurity) = DecisionTree::calculate_split(&X, 0, &y, &indices);
assert!(threshold == -0.5);
assert!(split_impurity == 0.0);

fn gini_impurity(left_child_proportion: f64, left_child_probability: f64, right_child_probability: f64) -> f64

Determine the information gain for a split during learning.

Arguments:

left_child_proportion - proportion of target data that map to left child data

left_child_probability - proportion of training data that contain left child data

right_child_probability - proportion of training data that contain right child data

Example:

let impurity = DecisionTree::gini_impurity(0.2, 1.0, 0.5);
let expected = 0.8 * 0.5;
assert!(impurity == expected);

fn build_tree(&mut self, X: &Mat<f64>, y: &Col<f64>, indices: &Col<usize>, depth: usize) -> Node

Build a decision tree on training data.

Arguments:

X - training data

y - target data

indices - indices of data subset fed to node

depth - current depth of tree

fn query_tree(&self, node: &Node, X: &Mat<f64>, row_idx: usize) -> f64

Traverse tree to retreive terminal classification probability on new data.

Arguments:

X - test data

y - target data

row_idx - row index of test sample in test data