Struct rsml::tree::model::DecisionTree

pub struct DecisionTree {
    pub max_depth: i32,
    pub min_samples_split: i32,
    pub n_features: usize,
    pub n_outputs: usize,
    pub classes: Col<f64>,
    pub n_classes: usize,
    pub used_features: Vec<usize>,
    pub root: Option<Node>,
}

DecisionTree represents the full decision tree model

Fields

max_depth

maximum depth of the tree

min_samples_split

minimum number of samples to split on

n_features

number of features

n_outputs

number of outputs

classes

array of classes

n_classes

number of classes

used_features

features split on

root

root node of tree

Methods

impl DecisionTree

fn new() -> DecisionTree

create new decision tree

fn split(feature: Feature<f64>, threshold: f64) -> (Vec<usize>, Vec<usize>)

Split a feature column in training data.

Arguments:

X - original training data

indices - indices of data subset fed to node

feature_idx - index of current feature to split on

threshold - threshold to apply to current split

Example:

extern crate rsml;
extern crate ndarray;
extern crate rand;
extern crate ndarray_rand;
use ndarray_rand::RandomExt;
use ndarray::RcArray;
use rand::distributions::Range;
use rsml::tree::model::DecisionTree;
fn main(){
let z = RcArray::random((10,5), Range::new(0.,10.));
let feature_idx = 4;
let value = 4.0;
let (left, right) = DecisionTree::split(z.column(feature_idx),value);
assert!(left.iter().all(|&x| z.get((x,feature_idx)).unwrap() <= &value));
assert!(right.iter().all(|&x| z.get((x,feature_idx)).unwrap() > &value))
}

fn find_optimal_split(feature: Feature<f64>, target: &Col<f64>) -> (f64, f64)

Determine optimal threshold to split data

Arguments:

X - original training data

feature_idx - index of current feature to split on

y - original target data

indices - indices of data subset fed to node

Example:

extern crate ndarray;
extern crate rsml;
use ndarray :: {rcarr2, RcArray};
use rsml::tree::model::*;
fn main(){
let x = rcarr2(&[[-1.0], [-0.5], [0.0], [0.0],[0.0],[0.5],[1.0]]);
let y = RcArray::from_vec(vec![1.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0]);
let (threshold, split_impurity) = DecisionTree::find_optimal_split(x.column(0), &y);
assert!(threshold == -0.5);
assert!(split_impurity == 0.0);
}

fn gini_impurity(left_child_proportion: f64, left_child_probability: f64, right_child_probability: f64) -> f64

Determine the information gain for a split during learning.

Arguments:

left_child_proportion - proportion of target data that map to left child data

left_child_probability - proportion of training data that contain left child data

right_child_probability - proportion of training data that contain right child data

Example:

extern crate rsml;
use rsml::tree::model::DecisionTree;
fn main(){
let impurity = DecisionTree::gini_impurity(0.2, 1.0, 0.5);
let expected = 0.8 * 0.5;
assert!(impurity == expected);
}

fn build_tree(&mut self, X: &Mat<f64>, y: &Col<f64>, indices: &Vec<usize>, depth: usize) -> Node

Build a decision tree on training data.

Arguments:

X - training data

y - target data

indices - indices of data subset fed to node

depth - current depth of tree

fn query_tree(&self, node: &Node, X: &Mat<f64>, row_idx: usize) -> f64

Traverse tree to retreive terminal classification probability on new data.

Arguments:

X - test data

y - target data

row_idx - row index of test sample in test data